Damage to players

+++ Assets/Scripts/PlayerWeapon.cs
@@ -5,7 +5,7 @@ public class PlayerWeapon
 {
     public string name = "Pistol";
 
-    public float damage = 10f;
+    public int damage = 10;
 
-    public float range = 200;
+    public float range = 200f;
 }
